+++ next.config.js
@@ -2,7 +2,14 @@
 const nextConfig = {
     experimental: {
         appDir: true
-    }
+    },
+    webpack: (config, { isServer }) => {
+        if (!isServer) {
+            config.resolve.fallback.fs = false
+        }
+
+        return config
+    },
 }
 
 module.exports = nextConfig
