+++ package.json
@@ -18,6 +18,7 @@
     "@mui/material": "^5.13.2",
     "@mui/x-date-pickers": "^6.7.0",
     "@tanstack/react-query": "^4.29.7",
+    "@types/file-saver": "^2.0.5",
     "@types/node": "20.2.3",
     "@types/react": "18.2.6",
     "@types/react-dom": "18.2.4",
@@ -27,7 +28,7 @@
     "eslint": "8.41.0",
     "eslint-config-next": "13.4.3",
     "file-saver": "^2.0.5",
-    "next": "13.4.3",
+    "next": "^13.4.10",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-hook-form": "^7.44.3",
@@ -35,7 +36,6 @@
     "sass": "^1.62.1",
     "typescript": "5.0.4",
     "uuid": "^9.0.0",
-    "zod": "^3.21.4",
-    "@types/file-saver": "^2.0.5"
+    "zod": "^3.21.4"
   }
 }
