+++ src/app/api/auth/[...auth0]/route.ts
@@ -1,0 +1,5 @@
+import {handleAuth, handleLogin} from '@auth0/nextjs-auth0';
+
+export const GET = handleAuth({
+    login: handleLogin({ authorizationParams: { audience: 'manager-api' } })
+});
