+++ src/app/snippets/test/[id]/page.tsx
@@ -1,9 +1,10 @@
-import {INITIAL_SNIPPETS, INITIAL_TEST} from "@/data/fake/fakeSnippetStore";
+import {INITIAL_TEST} from "@/data/fake/fakeSnippetStore";
 import TestSnippetPage from "@/modules/snippets/test/testSnippet";
+import {fetcher} from "@/data/fetcher";
+import {Snippet} from "@/data/snippet";
 
 const TestSnippet = async ({params}: { params: { id: string } }) => {
-    // const snippet = await fetcher<Snippet>(`manager/${params.id}`)
-    const snippet = INITIAL_SNIPPETS.find(s => s.id === params.id) ?? INITIAL_SNIPPETS[0]
+    const snippet = await fetcher<Snippet>(`manager/${params.id}`)
     return <TestSnippetPage snippet={snippet} tests={INITIAL_TEST}/>
 }
