+++ src/app/snippets/view/[id]/page.tsx
@@ -1,9 +1,10 @@
 import {INITIAL_SNIPPETS} from "@/data/fake/fakeSnippetStore";
 import ViewSnippetPage from "@/modules/snippets/view/viewSnippet";
+import {Snippet} from "@/data/snippet";
+import {fetcher} from "@/data/fetcher";
 
 const ViewSnippet = async ({params}: { params: { id: string } }) => {
-    // const snippet = await fetcher<Snippet>(`manager/${params.id}`)
-    const snippet = INITIAL_SNIPPETS.find(s => s.id === params.id) ?? INITIAL_SNIPPETS[0]
+    const snippet = await fetcher<Snippet>(`manager/snippet/${params.id}`)
     return <ViewSnippetPage snippet={snippet}/>
 }
