+++ src/data/snippet.ts
@@ -3,7 +3,7 @@ import {z} from 'zod'
 export const ComplianceSchema = z.enum(['pending', 'failed', 'not-compliant', 'compliant'])
 export type Compliance = z.infer<typeof ComplianceSchema>
 
-export const SnippetTypeSchema = z.enum(['printscript'])
+export const SnippetTypeSchema = z.enum(['PRINTSCRIPT'])
 export type SnippetType = z.infer<typeof SnippetTypeSchema>
 
 export const SnippetDescriptorSchema = z.object({
