+++ src/middleware.ts
@@ -1,15 +1,3 @@
-// middleware.js
-import { withMiddlewareAuthRequired, getSession } from '@auth0/nextjs-auth0/edge';
-import {NextResponse} from "next/server";
+import {withMiddlewareAuthRequired} from '@auth0/nextjs-auth0/edge';
 
-export default withMiddlewareAuthRequired(async function middleware(req) {
-    const res = NextResponse.next({
-        request: {
-            headers: new Headers(req.headers)
-        }
-    });
-
-    const user = await getSession(req, res);
-    res.cookies.set('hl', user?.language);
-    return res;
-});
+export default withMiddlewareAuthRequired();
